+++ IconicFramework/Constants.cs
@@ -2,11 +2,11 @@ namespace LeFauxMods.IconicFramework;
 
 internal static class Constants
 {
-    public const string DataPath = ModId + "/Data";
+    public const string DataPath = "furyx639.ToolbarIcons/Data";
 
-    public const string IconPath = ModId + "/Icons";
+    public const string IconPath = "furyx639.ToolbarIcons/Icons";
 
-    public const string ModId = "furyx639.ToolbarIcons";
+    public const string ModId = "furyx639.IconicFramework";
 
-    public const string UiPath = "/UI";
+    public const string UiPath = "furyx639.ToolbarIcons/UI";
 }