+++ IconicFramework/Integrations/GenericModConfigMenu.cs
@@ -1,3 +1,4 @@
+using LeFauxMods.Common.Integrations.GenericModConfigMenu;
 using LeFauxMods.Common.Integrations.IconicFramework;
 using LeFauxMods.IconicFramework.Utilities;
 using Microsoft.Xna.Framework;
@@ -8,21 +9,34 @@ namespace LeFauxMods.IconicFramework.Integrations;
 internal sealed class GenericModConfigMenu
 {
     private const string Id = "spacechase0.GenericModConfigMenu";
+    private readonly GenericModConfigMenuIntegration gmcm;
+    private readonly IManifest manifest;
+    private readonly IReflectedMethod? method;
 
     /// <summary>
     ///     Initializes a new instance of the <see cref="GenericModConfigMenu" /> class.
     /// </summary>
     /// <param name="api">The Iconic Framework API.</param>
+    /// <param name="gmcm">Dependency for integration with generic mod config menu.</param>
+    /// <param name="manifest">Dependency for accessing mod manifest.</param>
     /// <param name="reflection">Dependency used for reflecting into non-public code.</param>
-    public GenericModConfigMenu(IIconicFrameworkApi api, IReflectionHelper reflection)
+    public GenericModConfigMenu(IIconicFrameworkApi api, GenericModConfigMenuIntegration gmcm, IManifest manifest,
+        IReflectionHelper reflection)
     {
+        this.gmcm = gmcm;
+        this.manifest = manifest;
+        if (!gmcm.IsLoaded)
+        {
+            return;
+        }
+
         if (!IntegrationHelper.TryGetMod(Id, out var mod))
         {
             return;
         }
 
-        var method = reflection.GetMethod(mod, "OpenListMenu", false);
-        if (method is null)
+        this.method = reflection.GetMethod(mod, "OpenListMenu", false);
+        if (this.method is null)
         {
             return;
         }
@@ -33,13 +47,26 @@ internal sealed class GenericModConfigMenu
             new Rectangle(16, 0, 16, 16),
             I18n.Button_GenericModConfigMenu_Title,
             I18n.Button_GenericModConfigMenu_Description);
-        api.Subscribe(
-            e =>
-            {
-                if (e.Id == Id)
-                {
-                    method.Invoke(0);
-                }
-            });
+
+        api.Subscribe(this.OnIconPressedEventArgs);
+    }
+
+    private void OnIconPressedEventArgs(IIconPressedEventArgs e)
+    {
+        if (e.Id != Id || !this.gmcm.IsLoaded)
+        {
+            return;
+        }
+
+        switch (e.Button)
+        {
+            case SButton.MouseRight or SButton.ControllerB:
+                this.gmcm.Api.OpenModMenu(this.manifest);
+                return;
+
+            default:
+                this.method?.Invoke(0);
+                return;
+        }
     }
 }