+++ IconicFramework/ModEntry.cs
@@ -1,3 +1,4 @@
+using LeFauxMods.Common.Integrations.GenericModConfigMenu;
 using LeFauxMods.Common.Integrations.RadialMenu;
 using LeFauxMods.Common.Services;
 using LeFauxMods.Common.Utilities;
@@ -16,48 +17,54 @@ internal sealed class ModEntry : Mod
     private readonly Dictionary<string, IconComponent> icons = [];
     private ModConfig config = null!;
     private ConfigHelper<ModConfig> configHelper = null!;
+    private GenericModConfigMenuIntegration gmcm = null!;
 
     /// <inheritdoc />
     public override void Entry(IModHelper helper)
     {
         // Init
-        this.configHelper = new ConfigHelper<ModConfig>(this.Helper);
+        this.configHelper = new ConfigHelper<ModConfig>(helper);
         this.config = this.configHelper.Load();
-        _ = new ConfigMenu(
-            helper,
-            this.ModManifest,
-            this.config,
-            this.configHelper,
-            this.icons);
+        this.gmcm = new GenericModConfigMenuIntegration(this.ModManifest, helper.ModRegistry);
+        if (this.gmcm.IsLoaded)
+        {
+            _ = new ConfigMenu(
+                helper,
+                this.ModManifest,
+                this.config,
+                this.configHelper,
+                this.gmcm,
+                this.icons);
+        }
 
-        I18n.Init(this.Helper.Translation);
+        I18n.Init(helper.Translation);
         Log.Init(this.Monitor);
 
-        var themeHelper = ThemeHelper.Init(this.Helper);
-        themeHelper.AddAsset(Constants.IconPath, this.Helper.ModContent.Load<IRawTextureData>("assets/icons.png"));
-        themeHelper.AddAsset(Constants.UiPath, this.Helper.ModContent.Load<IRawTextureData>("assets/ui.png"));
+        var themeHelper = ThemeHelper.Init(helper);
+        themeHelper.AddAsset(Constants.IconPath, helper.ModContent.Load<IRawTextureData>("assets/icons.png"));
+        themeHelper.AddAsset(Constants.UiPath, helper.ModContent.Load<IRawTextureData>("assets/ui.png"));
 
         // Integrations
         var modInfo = this.Helper.ModRegistry.Get(this.ModManifest.UniqueID)!;
-        var api = new ModApi(modInfo, this.Helper, this.config, this.icons);
-        _ = new IntegrationHelper(this.Helper.ModRegistry, this.Helper.Reflection);
-        _ = new AlwaysScrollMap(api, this.Helper.Reflection);
+        var api = new ModApi(modInfo, helper, this.config, this.icons);
+        _ = new IntegrationHelper(helper.ModRegistry, this.Helper.Reflection);
+        _ = new AlwaysScrollMap(api, helper.Reflection);
         _ = new Calendar(api);
-        _ = new CjbCheatsMenu(api, this.Helper.Reflection);
-        _ = new CjbItemSpawner(api, this.Helper.Reflection);
-        _ = new ContentPack(this.Helper, api);
+        _ = new CjbCheatsMenu(api, helper.Reflection);
+        _ = new CjbItemSpawner(api, helper.Reflection);
+        _ = new ContentPatcher(helper, api);
         _ = new DailyQuests(api);
-        _ = new GenericModConfigMenu(api, this.Helper.Reflection);
-        _ = new SpecialOrders(api);
-        _ = new StardewAquarium(api, this.Helper.Reflection);
+        _ = new GenericModConfigMenu(api, this.gmcm, this.ModManifest, helper.Reflection);
+        _ = new SpecialOrders(api, helper);
+        _ = new StardewAquarium(api, helper.Reflection);
         _ = new ToDew(api);
         _ = new ToggleCollisions(api);
 
         // Events
-        this.Helper.Events.Content.AssetRequested += OnAssetRequested;
-        this.Helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
-        this.Helper.Events.GameLoop.ReturnedToTitle += OnReturnedToTitle;
-        this.Helper.Events.GameLoop.SaveLoaded += this.OnSaveLoaded;
+        helper.Events.Content.AssetRequested += OnAssetRequested;
+        helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
+        helper.Events.GameLoop.ReturnedToTitle += OnReturnedToTitle;
+        helper.Events.GameLoop.SaveLoaded += this.OnSaveLoaded;
     }
 
     /// <inheritdoc />
@@ -67,8 +74,7 @@ internal sealed class ModEntry : Mod
     {
         if (e.NameWithoutLocale.IsEquivalentTo(Constants.DataPath))
         {
-            e.LoadFrom(
-                static () => new Dictionary<string, ContentPackData>(StringComparer.OrdinalIgnoreCase),
+            e.LoadFrom(static () => new Dictionary<string, ContentData>(StringComparer.OrdinalIgnoreCase),
                 AssetLoadPriority.Exclusive);
         }
     }